"Refactor dialog open state and incorporate TypeScript
The dialog open state in main.tsx was initially set to a binary switch (open/close) which lacked context. This state was refactored to accept a boolean argument indicating the desired state instead therefore providing context. A TypeScript expect error was also added in the useEffect hook to handle a TypeScript issue wherein the setTimeout return value does not correctly match the expected 'Timeout' type."