Optimize focus trap and component refactoring
In this commit, the Main component was replaced with the MainClass in page.tsx and several refactoring changes were made on the latter for better code readability and maintainability. The component references were more explicitly named, adjusting from 'containerRef' and 'buttonRef' to 'modalContainerRef' and 'closeButtonRef' respectively for clearer understanding. The 'handleEnterRefFocus' method was also simplified, removing an unnecessary setTimeout function.

Furthermore, the focus trapping functionality for accessible modals was improved. The 'keydown' event listener in focus-trap.ts was simplified by removing the unnecessary variable 'KEYCODE_TAB'. This change was made because the 'Tab' key is being handled as a string, making the variable redundant.

In main.tsx, additional event listeners were added to improve accessibility features. A to-be-focussed button can now not only be triggered by the 'Enter' key but also by a mouse click. The same applies to the exit functionality, which can also be triggered using the 'Escape' key.

These changes were guided by the principles of improving Maintainability and User Experience.