Update focus-trap event handling and upgrade dependencies
This commit updates the handling of shift/tab key events in the React focus-trap component for better readability. Commented information was removed as it seemed unnecessary for understanding.
In the second part, the dependencies are updated particularly to eslint-config-ethang v8.0.2, '@turbo/gen' v1.10.12, and eslint v8.46.0. This update mainly includes upgrades to ESLint, TypeScript, and other related dependencies for improved static code analysis, type checking, and developer experience.