Add 'create post' functionality to api and UI
Added a new method 'createPost' to the api object for making POST requests. Adjusted the existing api routes and types to include new post creation. Created a new 'CreatePost' component and included it in the HomePage. Updated the existing PostsData component to utilize the new api object. Installed '@ethang/use-form' for handling form state in the new 'CreatePost' component. Add POST method to handle post creation in the backend and updated the data validation schema.