Improve search service by utilizing AuctionService API
In this change, the search service is now using an API from AuctionService. An AuctionServiceHttpClient is introduced to send requests to the API. The `DbInitializer` code is refactored to utilize the new client, replacing the old data loading mechanism. To support this, `appsettings.Development.json` is updated to contain `AuctionServiceUrl` and new Polly policy is added for the client to handle network failures.

These changes lead to a more reliable and scalable source of data for the search service, instead of merely relying on a local file. Also, it now only fetches updated data, reducing unnecessary traffic. Modifications in `AuctionsController` allow it to accept 'date' as parameter to filter the results.