Update IdentityService and AuctionService for Postman based authentication
Changes were made in the IdentityService and AuctionService to replace the previous authentication system based on 'm2m.client' and 'interactive' clients with a Postman based one. This was done to facilitate debugging and testing of secured APIs. Oauth2 resource owner password grant type was used for Postman client.

Addition of JWT bearer authentication in the AuctionService was done to implement claim-based authentication using the Postman client. User identity name is now used for all CRUD operations.

A new profile service 'CustomProfileService' was added in the IdentityService to generate claims for JWT tokens. Changes were also made in the nav bar and a new alternate text was added for better website accessibility.

The "IdentityServiceUrl" was added in appsettings.Development.json and Srting authentication has been added in Program.cs, this will be used for token generation and validation.

Bumped Microsoft.AspNetCore.Authentication.JwtBearer version to 7.0.10 in the AuctionService.