Add consumers for auction finished and bid placed events
This commit introduces two new consumers - AuctionFinishedConsumer and BidPlacedConsumer in the SearchService. These are used to handle events when an auction is finished or a new bid is placed respectively. The AuctionFinishedConsumer updates the auction's status to 'Finished' and saves the winner's information if the item is sold. The BidPlacedConsumer updates the current highest bid of the auction if the new bid exceeds the current bid and is accepted.