Add new client to IdentityService config
The IdentityService Config has been updated to add a new client named "nextApp". This was needed to facilitate the authorization of this new app, providing it with its own specific identity and grant types. The client is set to use Authorization Code and Client Credentials grant types, without PKCE requirement. It has a designated redirect URI and allowed scopes, and offline access is allowed. The access token lifetime is set to last for a month.