Refactor code and organize import statements
Restructured the order of variables and import in the classes within the IdentityService, AuctionService, GatewayService, and SearchService folders. All redundant braces and comments were removed and all .csproj files had their indentation fixed.

Having the import statements and variables ordered provides better readability to the code, hence it is best practice to keep the code and import statements well-ordered and organised. Also, removing unnecessary braces and comments enables developers to read and understand the code faster and more efficiently.