"Add GatewayService to docker-compose and refine configuration
This commit introduces multiple changes primarily aiming to add GatewayService to the docker-compose configuration and further fine-tune the application's settings. A new signing key for the identity service was specified in a new 'is-signing-key' file, enhancing the app's security. The docker-compose file now incorporates 'gateway-svc' into the service list, allowing it to work in conjunction with existing services. Debugging was improved in 'Index.cshtml.cs' file by expanding the list of local addresses. Essential settings and routes for GatewayService were introduced in 'appsettings.json' and 'appsettings.Development.json'. 'Dockerfile' and 'appsettings.Docker.json' were added to allow proper Dockerization and configuration of GatewayService. The changes would enable the GatewayService to direct HTTP requests to other services in a Docker environment effectively."