Add project setup & basic application layout
Changes include the configuration files for the project development, visible sections & initial functionality of the application, and a set of auxiliary files & scripts. The major frameworks/tools used are Next.js, TypeScript, Tailwind CSS, ESLint. Aimed to establish a solid project structure & workflow, plus an initial implementation of the application's UI/UX. Besides the configuration files, newly added components are for the auction listings feature, including a countdown timer and a card to display individual auction.