Add Pagination and Item selection features to the Auction App
This commit adds two main features to the Auction App.
First, pagination is introduced to break down the list of auctions into more user-friendly chunks
and improve the performance of the application by reducing the quantity of data to be handled at once.
Changes include the addition of a Filter component into the Listings component and modifying the getData
function to accept pageSize as an argument. The package.json file and pnpm-lock.yaml files were updated to
include the @nextui-org/select package to support this feature.

Secondly, it introduces the ability for the user to change the number of items displayed per page.
This gives users more flexibility in their viewing preferences, and allows them to optimize the view according
to their screen size or data plan.

To enable this second feature, the Filter component in filter.tsx was updated to include
selection options (4, 8 and 12 items per page), and the listings.tsx was modified to include pageSize state
and setPageSize function for managing the number of items shown per page. The getData function in auction-actions.ts
is also updated to fetch data according to the selected page size.

In addition, this commit includes commits to .idea/git_toolbox_prj.xml for the project settings.