Fix data loading check in auction listings
Removed a console.log from the auction actions that was outputting the search parameters for debugging. Also, updated the check for data in the auctions listing file to use isNil from lodash to prevent potential issues with false positives when checking for data. This was possibly causing some erroneous renderings in the auction listings.