Add auction updates and related functionalities
This commit introduces updates to the auction based features of our application. New functionalities are added to allow users to create an auction, view the details of a selected auction, and list their won or created auctions. Specifically, `AuctionDetails` and `UpdateAuction` components were added to allow users to interact with specific auctions.
The user interface was also updated to enhance the functionality. Buttons were added under the user actions dropdown for 'My Auctions' and 'Auctions Won', these buttons redirect the user to the respective views. The `useParameterStore` was updated to include seller and winner of the auction, this helps in separating out the auctions won and created by the user.
Also, the loading state was added to the `listings` to improve UX during data fetching. Minor updates in the `auth` route and `auction-actions` were made for better token management and data fetching. Additional dependencies were added for form management.
This update enhances the usability of our application by providing users with an intuitive way to manage and interact with auctions.