Add Bidding Service
Integrated a bidding service component, which is responsible for management of bids for an auction. This includes receiving, processing, logging, and interacting with auction and identity services. It also includes some correlation logic to decide whether a bid should be accepted or not. A Docker configuration is also added to make the service ready for containerization. Key technologies used in this addition are .NET SDK, RabbitMQ for message transportation, MongoDB for bid persistence and JWT bearer for Authentication.

Additional files for BiddingService are added in the src folder structure and the necessary references are added to the solution (Carsties.sln) and Docker files of other services. A new model for auctions has been added for local tracking of auctions with properties including an ID, Seller name, Auction end time, Reserved price, and auction state.

Other major inclusions are Controllers for processing requests, DTOs for mapping request and response data, MappingProfiles for object-to-object mapping, and Consumers for processing in-queue messages. Necessary configurations are done for localhost testing. All these are to ensure layered communication between components and separation of concerns within the Bidding Service.

The docker-compose.yml has got a new service definition for bidding-svc detailing the build configuration, environment variables, and dependent services.

Necessary changes are made in the Auction service to parse the auction id correctly, ensuring integer to GUID conversion for compatibility with the newly included Bidding Service.