Remove Roman to Integer and Valid Parentheses functions
Deleted the 'leet-code' folder that contained the Roman to Integer ('romanToInt') and the Valid Parentheses ('isValidParentheses') functions, along with their respective test files and package.json. The decision to delete these files was made as they were no longer needed in the current context of the project. The changes also include the removal of the dependencies used by these two functions.