Create package.json for algorithms application
This commit adds a new package.json file to the algorithms directory. This is required to manage local dependencies for our algorithm features. The file outlines the main entry point for the app as 'index.js' and also defines important scripts for testing using 'vitest'. It includes references to necessary dependencies like '@faker-js/faker', '@types/node', 'eslint-config-ethang', 'typescript', and 'vitest'.