Add 'first non-repeating character' algorithm and tests
Added the 'first non-repeating character' algorithm in two formats: one using a basic object and another using a map. These functions find the first unique character in a given string. Also added corresponding tests to verify that these functions are working as expected, and a benchmark test, utilizing a random string generator, to measure the performance of both implementations.