Add IPv4 validation functions and tests
This commit adds two new files in the 'is-valid-ipv4' directory under 'algorithms' app. The first file, 'is-valid-ipv4.ts', includes two functions 'isValidIpv4' and 'isValidIpv4Regex' for validating if a given string follows the IPv4 address standard - either through a split and check method or a regular expression. The second file 'is-valid-ipv4.test.ts' includes unit tests for these validation functions with predefined test cases. Also, a benchmarking script 'is-valid-ipv4.bench.ts' has been added for performance testing of the functions.