Improve ES6 js immutability and create lint script
This commit aims to improve the ES6 immutability of the javascript files by replacing variables with mutable declarations (using "let") and spread operator whenever possible. An eslint script command was added to make it easier to check and fix for linting issues. This change was made for enforcing best practices in fututre code addition/modification and maintaining code quality throughout the project. Several files including algorithms and test cases were modified for this purpose.