Add Queue, Stack, and LinkedList classes
Added Queue, Stack, and LinkedList classes for their use in algorithms. These data structures are fundamental for many algorithm tasks. Added methods for each data structure to add, remove, check if it's empty, and get the length. Applied these new data structures in isPalindrome algorithm for new isPalindromeQueue and isPalindromeStack methods. Also created is the balancedParentheses that checks for balanced parentheses in a string using stack data structure.

Additionally, the eslint-config-ethang package version was updated for improved linting. Test files were also updated to include tests for these modifications.

This addition improves the capability of the algorithms that can be implemented and tested using Queue, Stack, and LinkedList. The eslint-config-ethang package update ensures code standard is maintained, and the updated tests ensure the functionality of the newly implemented code.