Update @ethang/project-builder version to 1.1.22
The version of @ethang/project-builder in the project dependencies has been updated from 1.1.21 to 1.1.22 as specified in both pnpm-lock.yaml and package.json files. This helps to ensure our project utilizes the most recent version, benefiting from the latest features, and maintains a consistency across all versions in the project.